[TabBar] Refactor dividers to a separate file and optimize style.
